FLTK library updated

+++ fltk_win32/share/doc/fltk/examples/CubeView.cxx
@@ -1,24 +1,15 @@
 //
-// "$Id: CubeView.cxx 7903 2010-11-28 21:06:39Z matt $"
+// "$Id: CubeView.cxx 9325 2012-04-05 05:12:30Z fabien $"
 //
 // CubeView class implementation for the Fast Light Tool Kit (FLTK).
 //
 // Copyright 1998-2010 by Bill Spitzak and others.
 //
-// This library is free software; you can redistribute it and/or
-// modify it under the terms of the GNU Library General Public
-// License as published by the Free Software Foundation; either
-// version 2 of the License, or (at your option) any later version.
+// This library is free software. Distribution and use rights are outlined in
+// the file "COPYING" which should have been included with this file.  If this
+// file is missing or damaged, see the license at:
 //
-// This library is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-// Library General Public License for more details.
-//
-// You should have received a copy of the GNU Library General Public
-// License along with this library; if not, write to the Free Software
-// Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA 02111-1307
-// USA.
+//     http://www.fltk.org/COPYING.php
 //
 // Please report all bugs and problems on the following page:
 //
@@ -158,8 +149,8 @@ void CubeView::draw() {
 
     glPushMatrix();
 
-    glTranslatef(xshift, yshift, 0);
-    glRotatef(hAng,0,1,0); glRotatef(vAng,1,0,0);
+    glTranslatef((GLfloat) xshift, (GLfloat) yshift, 0);
+    glRotatef((GLfloat) hAng,0,1,0); glRotatef((GLfloat) vAng,1,0,0);
     glScalef(float(size),float(size),float(size));
 
     drawCube();
@@ -169,5 +160,5 @@ void CubeView::draw() {
 #endif /* HAVE_GL */
 
 //
-// End of "$Id: CubeView.cxx 7903 2010-11-28 21:06:39Z matt $".
+// End of "$Id: CubeView.cxx 9325 2012-04-05 05:12:30Z fabien $".
 //
